refactor: optimize useWalletTag
